feat(referral): fix issue with temp ref API endpoint
Fixes an issue where the `getTempRef` function was not properly
constructing the API endpoint URL. The `address` parameter was not
being properly interpolated into the URL, causing the API request to
fail. This change fixes the URL construction to ensure the `address`
parameter is correctly included in the API endpoint.