feat: disable rendering when no stakers
The changes in this commit disable the rendering of the `LevelOneNewStakeModal` component when there are no stakers. This is done by commenting out the `setRender(!!0)` line in the `useEffect` hook. This change ensures that the modal is not displayed when there are no stakers, improving the user experience.

Additionally, the commit updates the `API_URL` configuration in the `config/index.ts` file. The previous URL for the "prev" environment is replaced with the correct URL for the Nestage API.